Grove: use Seeed rainbow and isolate I/O code
